+++ lldb/source/DataFormatters/LibCxxList.cpp
@@ -263,7 +263,7 @@ lldb_private::formatters::LibcxxStdListSyntheticFrontEnd::GetChildAtIndex (size_
         return lldb::ValueObjectSP();
     
     StreamString name;
-    name.Printf("[%zu]",idx);
+    name.Printf("[%" PRIu64 "]", (uint64_t)idx);
     return (m_children[idx] = ValueObject::CreateValueObjectFromData(name.GetData(), data, m_backend.GetExecutionContextRef(), m_element_type));
 }
